continue work with CRUD

+++ ShopAction.Data/Entities/ProductImage.cs
@@ -12,7 +12,7 @@ namespace ShopAction.Data.Entities
         public bool IsDefault { get; set; }
         public DateTime DateCreated { get; set; }
         public int SortOrder { get; set; }
-        public int FileSize { get; set; }
+        public long FileSize { get; set; }
         public Product Product { get; set; }
     }
 }
